cosmetic changes added custom conf file
