+++ modules/sc-mesh-secure-deployment/src/1.5/common/mesh_com_11s.conf
@@ -1,0 +1,26 @@
+---
+debug: True
+client:
+    set_hostname: True
+    disable_networking: True
+    mesh_service: True
+    gw_service: True
+    dflt_service: True
+    mesh_inf: 'wlp'
+    gw_inf: 'wla'
+server:
+    mesh_inf: 'wl'
+    secos:
+      api_version: 1                  # interface version for future purposes
+      ssid: "gold"                    # 0-32 octets, UTF-8, shlex.quote chars limiting
+      key: ""                         # key for the network
+      enc: "SAE"                      # encryption (none wep, wpa2, wpa3, sae)
+      ap_mac: "00:11:22:33:44:55"     # bssid for mesh network
+      country: "AE"                   # Country code, sets tx power limits and supported channels
+      frequency: 5180                 # 5180 wifi channel frequency, depends on the country code and HW
+      subnet: 255.255.255.0           # subnet mask
+      tx_power: 30                    # select 30dBm, HW and regulations limiting it correct level. Can be used to set lower dBm levels for testing purposes (e.g. 5dBm)
+      mode: "mesh"                    # mesh=mesh network, ap=debug hotspot
+      type: "11s"                     # 11s or ibss
+      ip: 10.10.10.4                  # Mesh IP address
+
