--- modules/sc-mesh-secure-deployment/src/1.5/features/mutual/mutual2.py
@@ -1,250 +1,0 @@
-#!/bin/python3
-
-import os
-from os import path
-from utils import wifi_ssrc as wf
-from utils import funsocket as fs
-import time
-import pandas as pd
-from utils import primitives as pri
-from termcolor import colored
-import sys
-
-sys.path.insert(0, '../../')
-'''
-only for testing 
-'''
-from common import ConnectionMgr
-
-co = ConnectionMgr.ConnectionMgr()
-
-
-def folder():
-    if not path.isdir(PATH):
-        os.mkdir(PATH)
-
-
-# always install it with python3 gnupg, do not use pip3 install gnupg (different libraries)
-PATH = '../auth'
-folder()
-
-root_cert = "/etc/ssl/certs/root_cert.der"
-local_cert = "/etc/ssl/certs/mesh_cert.der"
-
-try:
-    open(local_cert, 'rb')
-except FileNotFoundError:
-    print("Local certificate not found. Run generate_keys.sh")
-    sys.exit(0)
-try:
-    open(root_cert, 'rb')
-except FileNotFoundError:
-    print("Root certificate not found. Run generate_keys.sh")
-    sys.exit(0)
-
-
-class Mutual:
-    """
-    set_level function to establish trust level of an imported key
-    run is the state machine of the entire class
-    interface should be the wireless interface of the mutual AP.
-    """
-
-    def __init__(self, interface):
-        """
-        Import the keys: node Pub,Priv and server pub.
-        """
-        self.interface = interface
-        print('Loading keys')
-        self.table = self.create_table()
-        self.root_cert = root_cert
-        self.local_cert = local_cert
-        self.debug = True
-        self.cli = False
-        self.myID = pri.get_labels()
-        print("loading root_cert")
-        pri.import_cert(root_cert, 'root')
-        print("my ID: ", self.myID)
-
-    def create_table(self):
-        '''
-        Function to create table of authenticated devices
-        '''
-        columns = ['ID', 'MAC', 'IP', 'PubKey_fpr', 'MA_level']
-        if not path.isfile('../auth/dev.csv'):
-            table = pd.DataFrame(columns=columns)
-            table.to_csv('../auth/dev.csv', header=columns, index=False)
-        else:
-            table = pd.read_csv('../auth/dev.csv')
-        return table
-
-    def update_table(self, info):
-        '''
-        this function update the table with the node's info
-        '''
-        if info['ID'] not in set(self.table['ID']):
-            self.table = self.table.append(info, ignore_index=True)
-            self.table.to_csv('../auth/dev.csv', mode='a', header=False, index=False)
-        elif self.table.loc[self.table['ID'] == info['ID']]['PubKey_fpr'].all() != info['PubKey_fpr']:
-            self.table = self.table.append(info, ignore_index=True)
-            self.table.to_csv('../auth/dev.csv', mode='a', header=False, index=False)
-
-    def get_status(self):
-        '''
-        status of the current authenticated nodes
-        '''
-        return self.table.to_dict()
-
-    def signature(self):
-        _, sig = pri.hashSig(self.root_cert)
-        if self.debug:
-            print(f'> Signature: {bytes(sig)}')
-        return bytes(sig)
-
-    def digest(self):
-        dig, _ = pri.hashSig(self.root_cert)
-        if self.debug:
-            print(f'> Digest: {dig}')
-        return dig
-
-    def exchange(self, candidate, serverIP):
-        cert = open(self.local_cert, 'rb').read()
-        message = self.signature() + cert + self.myID.encode('utf-8')
-        fs.client_auth(candidate, serverIP, message)
-
-    def client(self, candidate):
-        print(f'AuthAP available: {candidate}')
-        wf.connect_wifi(candidate, self.interface)
-        print('2) sending signature of root certificate')
-        serverIP = (".".join(co.get_ip_address(self.interface).split(".")[:3]) + ".1")
-        self.exchange(candidate, serverIP)
-        print('3) getting node certificate')
-        server_sig, addr = fs.server_auth(self.myID, self.interface)
-        self.cli = True
-        return server_sig, addr
-
-    def server(self):
-        print('No AP available')
-        print("I'm an authentication server. Creating AP")
-        wf.create_ap(self.myID, self.interface)  # create AuthAPnodeID for authentication
-        time.sleep(2)
-        print('1) server default')
-        client_cert, addr = fs.server_auth(self.myID, self.interface)
-        return client_cert, addr
-
-    def decode_cert(self, client):
-        sig = self.signature()
-        client_sig = client[:len(sig)]
-        client_cert = client[len(sig):]
-        cliID = client[-5:]
-        return client_sig, client_cert, cliID.decode('utf-8')
-
-    def set_password(self, node_name):
-        password = co.get_password()
-        if password == '':  # this means still not connected with anyone
-            password = co.create_password()  # create a password (only if no mesh exist)
-            co.util.update_mesh_password(password)  # update password in config file
-        print(password)
-        encrypt_pass = pri.encrypt_response(password, node_name)
-        if self.debug:
-            print(f'encrypted pass: {str(encrypt_pass)}')
-        return encrypt_pass
-
-    def send_password(self, cliID, addr, my_mac_mesh, my_ip_mesh, encrypt_pass):
-        try:
-            time.sleep(2)
-            fs.client_auth(cliID, addr[0], encrypt_pass)  # send mesh password
-        except ConnectionRefusedError:
-            time.sleep(7)
-            fs.client_auth(cliID, addr[0], encrypt_pass)
-        client_mesh_ip, _ = fs.server_auth(self.myID, self.interface)
-        if self.debug:
-            print(f'Client Mesh IP: {str(client_mesh_ip)}')
-        try:
-            time.sleep(2)
-            fs.client_auth(cliID, addr[0], my_mac_mesh.encode())  # send my mac
-        except ConnectionRefusedError:
-            time.sleep(7)
-            fs.client_auth(cliID, addr[0], my_mac_mesh.encode())
-        client_mac, _ = fs.server_auth(self.myID, self.interface)
-        if self.debug:
-            print(f'Client Mesh MAC: {str(client_mesh_ip)}')
-        try:
-            fs.client_auth(cliID, addr[0], my_ip_mesh.encode())  # send my mesh ip
-        except ConnectionRefusedError:
-            time.sleep(7)
-            fs.client_auth(cliID, addr[0], my_ip_mesh.encode())
-        return client_mac, client_mesh_ip
-
-    def start_mesh(self):
-        print('creating mesh')
-        self.my_ip_mesh, self.my_mac_mesh = co.create_mesh_config()
-        co.start_mesh()
-
-    def start(self):
-        set_mesh = False
-        candidate = wf.scan_wifi(self.interface)  # scan wifi to authenticate with
-        if candidate:  # client
-            sigs, addr = self.client(candidate)
-            sig, client_cert, cliID = self.decode_cert(sigs)
-            cli = True
-        else:  # server
-            client, addr = self.server()
-            sig, client_cert, cliID = self.decode_cert(client)
-            print("Client ID", cliID)
-            print('4) local key')
-            self.exchange(cliID, addr[0])
-            cli = False
-        node_name = addr[0].replace('.', '_')
-        pri.import_cert(client_cert, node_name)
-        if pri.verify_certificate(sig, node_name, self.digest(), self.root_cert):
-            print(colored('> Valid Certificate', 'green'))
-            print('Authenticated, now send my pubkey')
-            client_fpr, _ = pri.hashSig(node_name + '.der')
-            if cli:  # client
-                print('5) get password')
-                enc_pass, _ = fs.server_auth(self.myID, self.interface)
-                password = pri.decrypt_response(enc_pass)
-                print(bytes(password).decode())
-                co.util.update_mesh_password(bytes(password).decode())
-                self.start_mesh()
-                try:
-                    fs.client_auth(cliID, addr[0], self.my_ip_mesh.encode())  # send my mesh ip
-                except ConnectionRefusedError:
-                    time.sleep(2)
-                    fs.client_auth(cliID, addr[0], self.my_ip_mesh.encode())
-                client_mac, _ = fs.server_auth(self.myID, self.interface)
-                print(f'Client MAC: {str(client_mac)}')
-                try:
-                    time.sleep(2)
-                    fs.client_auth(cliID, addr[0], self.my_mac_mesh.encode())  # send my mac
-                except ConnectionRefusedError:
-                    time.sleep(2)
-                    fs.client_auth(cliID, addr[0], self.my_mac_mesh.encode())
-                client_mesh_ip, _ = fs.server_auth(self.myID, self.interface)
-            elif pri.verify_certificate(sig, node_name, self.digest(), self.root_cert):
-                print(colored('> Valid Certificate', 'green'))
-                print("4.1) Setting PasswordS")
-                encrypt_pass = self.set_password(node_name)
-                self.start_mesh()
-                client_mac, client_mesh_ip = self.send_password(cliID, addr, self.my_mac_mesh, self.my_ip_mesh,
-                                                                encrypt_pass)
-            info = {'ID': node_name, 'MAC': client_mac.decode(), 'IP': client_mesh_ip.decode(),
-                    'PubKey_fpr': client_fpr, 'MA_level': 'A'}
-            self.update_table(info)  # #update csv file
-        else:
-            info = {'ID': node_name, 'MAC': "00:00:00:00", 'IP': addr[0],
-                    'PubKey_fpr': "___", 'MA_level': 'NA'}
-            self.update_table(info)  # #update csv file
-            print(colored("Not Valid Client Certificate", 'red'))
-            pri.delete_key(node_name)
-            os.remove(node_name + '.der')
-
-    def test(self): # unit test
-        raise NotImplementedError
-
-
-if __name__ == "__main__":
-    mutual = Mutual('wlan1')
-    # mutual.terminate()
-    mutual.start()
